+++ src/lib.rs
@@ -27,7 +27,7 @@ struct Foo {
     b: u32,
 }
 
-println!("{}", Foo::layout());
+println!("{}", Foo::type_layout());
 // prints:
 // Foo (size 8, alignment 4)
 // | Offset | Name      | Size |
@@ -49,7 +49,7 @@ struct OverAligned {
     value: u8,
 }
 
-println!("{}", OverAligned::layout());
+println!("{}", OverAligned::type_layout());
 // prints:
 // OverAligned (size 128, alignment 128)
 // | Offset | Name      | Size |
@@ -75,7 +75,7 @@ pub use type_layout_derive::TypeLayout;
 pub use memoffset;
 
 pub trait TypeLayout {
-    fn layout() -> TypeLayoutInfo;
+    fn type_layout() -> TypeLayoutInfo;
 }
 
 #[derive(Debug)]