+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Added `serde1` feature, which enables Serde serialization and deserialization for the crate's types.
+* Renamed `TypeLayout::layout` to `TypeLayout::type_layout` to be more unique.
 
 ## 0.1.0 (2020-09-12)
 * Initial MVP release